Moved imglib2 components to ome.scifio.img package
Moved the IO components from imglib2 to ome.scifio.img. Package
declarations updated.
